Flytt validering til egen service + fjern kall til prevalideringssteget til behandlingsresultatsteg
