+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/beregning/TilkjentYtelseValideringService.kt
@@ -1,0 +1,37 @@
+package no.nav.familie.ba.sak.kjerne.beregning
+
+import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
+import no.nav.familie.ba.sak.kjerne.grunnlag.personopplysninger.PersongrunnlagService
+import no.nav.familie.ba.sak.kjerne.totrinnskontroll.TotrinnskontrollService
+import org.springframework.stereotype.Service
+
+@Service
+class TilkjentYtelseValideringService(
+    private val totrinnskontrollService: TotrinnskontrollService,
+    private val beregningService: BeregningService,
+    private val persongrunnlagService: PersongrunnlagService
+) {
+    fun validerAtIngenUtbetalingerOverstiger100Prosent(behandling: Behandling) {
+        if (behandling.erMigrering()) return
+        val totrinnskontroll = totrinnskontrollService.hentAktivForBehandling(behandling.id)
+
+        if (totrinnskontroll?.godkjent == true) {
+            val tilkjentYtelse = beregningService.hentTilkjentYtelseForBehandling(behandlingId = behandling.id)
+
+            val personopplysningGrunnlag = persongrunnlagService.hentAktivThrows(behandlingId = behandling.id)
+
+            val barnMedAndreRelevanteTilkjentYtelser = personopplysningGrunnlag.barna.map {
+                Pair(
+                    it,
+                    beregningService.hentRelevanteTilkjentYtelserForBarn(it.aktør, behandling.fagsak.id)
+                )
+            }
+
+            TilkjentYtelseValidering.validerAtBarnIkkeFårFlereUtbetalingerSammePeriode(
+                behandlendeBehandlingTilkjentYtelse = tilkjentYtelse,
+                barnMedAndreRelevanteTilkjentYtelser = barnMedAndreRelevanteTilkjentYtelser,
+                personopplysningGrunnlag = personopplysningGrunnlag,
+            )
+        }
+    }
+}
